Move whisper/model.rb -> whisper/model/uri.rb
