+++ bindings/ruby/lib/whisper/model/uri.rb
